requirements.txt
torch
torchvision
torchaudio
kagglehub
numpy
pandas
matplotlib
plotly
opencv-python
scikit-learn
tqdm
ipython
nbformat
flask
flask-cors
pillow
